quota: move cluster resource quota helpers closer to usage
